requirements.txt
pygame==2.1.0
pygame-menu==4.2.2
Pillow==9.0.0